requirements.txt
discord.py
tls-client
aiohttp 
